Test workspace schemes
